Set up sane example only configuration
Server, Admin etc information must of course still be adapted to the respective network as other aspects as well. This was done so some more sane defaults can be applied (e.g. adding the referenced yet missing opers and servers classes) and so defaults no longer specify DALnet to prevent impersonation and confusion.